requirements.txt
Flask==3.0.3
numpy==2.0.0
opencv_python==4.10.0.84
opencv_python_headless==4.10.0.84
Pillow==10.3.0
selenium==4.22.0
